WidgetButtons Refactor
