+++ src/types/icons.ts
@@ -1,0 +1,40 @@
+export type IconName =
+    | 'plus'
+    | 'copy'
+    | 'trash'
+    | 'times'
+    | 'check'
+    | 'plug'
+    | 'pencil'
+    | 'history'
+    | 'cross-circle'
+    | 'cross'
+    | 'upload'
+    | 'download'
+    | 'group'
+    | 'user'
+    | 'cubes'
+    | 'retweet'
+    | 'minus-square'
+    | 'info'
+    | 'gavel'
+    | 'push'
+    | 'sync'
+    | 'minus'
+    | 'lock'
+    | 'refresh'
+    | 'reload'
+    | 'handshake'
+    | 'compromise'
+    | 'destroy'
+    | 'bomb'
+    | 'search'
+    | 'random'
+    | 'sign'
+    | 'verify'
+    | 'key'
+    | 'link'
+    | 'recycle'
+    | 'rekey'
+    | 'archive'
+    | 'unarchive';
