+++ src/utils/icons.ts
@@ -1,0 +1,82 @@
+import React from 'react';
+import {
+    Plus,
+    Copy,
+    Trash2,
+    X,
+    Check,
+    Plug,
+    History,
+    XCircle,
+    Upload,
+    ArrowDownToLine,
+    Users,
+    User,
+    Boxes,
+    Repeat2,
+    MinusSquare,
+    Info,
+    Gavel,
+    ArrowUpCircle,
+    RefreshCw,
+    Minus,
+    Lock,
+    RotateCw,
+    Handshake,
+    Ban,
+    Bomb,
+    Search,
+    Shuffle,
+    Pen,
+    CheckSquare,
+    KeyRound,
+    Link,
+    Recycle,
+    Archive,
+    PackageOpen,
+    SquareMinus,
+} from 'lucide-react';
+import EditIcon from 'components/icons/EditIcon';
+import type { IconName } from 'types/icons';
+
+export const iconRegistry: Record<IconName, React.ComponentType<{ size?: number }>> = {
+    plus: Plus,
+    copy: Copy,
+    trash: Trash2,
+    times: X,
+    check: Check,
+    plug: Plug,
+    pencil: EditIcon,
+    history: History,
+    'cross-circle': XCircle,
+    cross: X,
+    upload: Upload,
+    download: ArrowDownToLine,
+    group: Users,
+    user: User,
+    cubes: Boxes,
+    retweet: Repeat2,
+    'minus-square': MinusSquare,
+    push: ArrowUpCircle,
+    sync: Repeat2,
+    info: Info,
+    minus: Minus,
+    gavel: Gavel,
+    lock: Lock,
+    refresh: RefreshCw,
+    reload: RotateCw,
+    handshake: Handshake,
+    compromise: Ban,
+    destroy: SquareMinus,
+    bomb: Bomb,
+    search: Search,
+    random: Shuffle,
+    sign: Pen,
+    verify: CheckSquare,
+    key: KeyRound,
+    link: Link,
+    recycle: Recycle,
+    rekey: Shuffle,
+    archive: Archive,
+    unarchive: PackageOpen,
+};
